+++ Doc/whatsnew/3.11.rst
@@ -572,7 +572,8 @@ Porting to Python 3.11
   header provides functions like ``printf()`` and ``fopen()``.
   (Contributed by Victor Stinner in :issue:`45434`.)
 
-* The non-limited API files ``cellobject.h`` and ``funcobject.h`` have been
+* The non-limited API files ``cellobject.h``, ``classobject.h``, ``context.h``,
+  ``funcobject.h``, ``genobject.h`` and ``longintrepr.h`` have been
   moved to the ``Include/cpython`` directory. These files must not be included
   directly, as they are already included in ``Python.h``: :ref:`Include Files
   <api-includes>`. If they have been included directly, consider including
