+++ Include/Python.h
@@ -46,7 +46,7 @@
 #include "bytesobject.h"
 #include "unicodeobject.h"
 #include "longobject.h"
-#include "longintrepr.h"
+#include "cpython/longintrepr.h"
 #include "boolobject.h"
 #include "floatobject.h"
 #include "complexobject.h"
@@ -61,7 +61,7 @@
 #include "methodobject.h"
 #include "moduleobject.h"
 #include "cpython/funcobject.h"
-#include "classobject.h"
+#include "cpython/classobject.h"
 #include "fileobject.h"
 #include "pycapsule.h"
 #include "code.h"
@@ -70,7 +70,8 @@
 #include "sliceobject.h"
 #include "cpython/cellobject.h"
 #include "iterobject.h"
-#include "genobject.h"
+#include "pystate.h"
+#include "cpython/genobject.h"
 #include "descrobject.h"
 #include "genericaliasobject.h"
 #include "warnings.h"
@@ -83,8 +84,7 @@
 #include "pyerrors.h"
 #include "cpython/initconfig.h"
 #include "pythread.h"
-#include "pystate.h"
-#include "context.h"
+#include "cpython/context.h"
 #include "modsupport.h"
 #include "compile.h"
 #include "pythonrun.h"
