+++ Misc/NEWS.d/next/C API/2021-10-15-00-11-51.bpo-35134.eX4zqy.rst
@@ -1,3 +1,4 @@
-Move ``cellobject.h`` and ``funcobject.h`` header files from ``Include/`` to
+Move ``cellobject.h``, ``classobject.h``, ``context.h``, ``funcobject.h``,
+``genobject.h`` and ``longintrepr.h`` header files from ``Include/`` to
 ``Include/cpython/``. C extensions should only include the main ``<Python.h>``
 header. Patch by Victor Stinner.
